Messing with the brush editor: Updating the nibs to xibs. Replacing deprecated functions.

+++ Brushed/BrushedDocumentController.h
@@ -19,6 +19,6 @@
 }
 
 // Allow customization of the open panel
-- (int)runModalOpenPanel:(NSOpenPanel *)openPanel forTypes:(NSArray *)extensions;
+- (NSInteger)runModalOpenPanel:(NSOpenPanel *)openPanel forTypes:(NSArray *)extensions;
 
 @end
